整理

+++ firmware/src/main.c
@@ -48,15 +48,14 @@ void ReadKeys()
   static uint8_t prev = 0;
   uint8_t curr = 0;
   uint8_t hold = 0, pressed = 0;
-  uint8_t keys =
-      (PORTAbits.RA0 << 0) |
-      (PORTAbits.RA1 << 1) |
-      (PORTAbits.RA3 << 2) |
-      (PORTAbits.RA4 << 3) |
-      (PORTBbits.RB4 << 4) |
-      (PORTBbits.RB5 << 5) |
-      (PORTBbits.RB6 << 6) |
-      (PORTBbits.RB7 << 7);
+  uint8_t keys = (uint8_t)((PORTAbits.RA0 << 0) |
+                           (PORTAbits.RA1 << 1) |
+                           (PORTAbits.RA3 << 2) |
+                           (PORTAbits.RA4 << 3) |
+                           (PORTBbits.RB4 << 4) |
+                           (PORTBbits.RB5 << 5) |
+                           (PORTBbits.RB6 << 6) |
+                           (PORTBbits.RB7 << 7));
 
   for (uint8_t n = 0; n < 8; n++)
   {
